+++ lib/src/fab/scroll_fab.dart
@@ -1,6 +1,7 @@
-import 'package:flutter_ahlib/list/scroll_fab_controller.dart';
+import 'package:flutter_ahlib/src/fab/scroll_fab_controller.dart';
 import 'package:flutter/material.dart';
 
+/// A fab which include scroll animation, which use `ScrollController` and register `ScrollFabController`
 class ScrollFloatingActionButton extends StatefulWidget {
   const ScrollFloatingActionButton({
     Key key,