+++ lib/src/list/refreshable_listview.dart
@@ -1,62 +1,53 @@
+import 'package:flutter_ahlib/src/list/placeholder_text.dart';
+import 'package:flutter_ahlib/src/list/scroll_more_controller.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_ahlib/list/placeholder_text.dart';
-import 'package:flutter_ahlib/list/scroll_more_controller.dart';
-import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
+import 'package:flutter_ahlib/src/list/type.dart';
 
-typedef GetNonPageDataFunction<T> = Future<List<T>> Function();
-
-class RefreshableStaggeredGridView<T> extends StatefulWidget {
-  const RefreshableStaggeredGridView({
+/// refreshable LisrView which packing `RefreshIndicator`, `ListPlaceholderText`, `Scrollbar` and `ListView`
+class RefreshableListView<T> extends StatefulWidget {
+  const RefreshableListView({
     Key key,
     @required this.data,
     @required this.getData,
     this.onStateChanged,
     this.placeholderText,
     this.padding,
+    this.separator,
     @required this.itemBuilder,
     this.controller,
     this.topWidget,
     this.bottomWidget,
-    this.primary,
-    this.crossAxisSpacing = 0,
-    this.mainAxisSpacing = 0,
-    @required this.crossAxisCount,
-    @required this.staggeredTileBuilder,
   })  : assert(data != null),
         assert(getData != null),
         assert(itemBuilder != null),
-        assert(crossAxisCount != null),
-        assert(staggeredTileBuilder != null),
         super(key: key);
 
   final List<T> data;
   final GetNonPageDataFunction<T> getData;
   final StateChangedCallback onStateChanged;
-  final ListPlaceHolderSetting placeholderText;
+  final ListPlaceholderSetting placeholderText;
   final EdgeInsetsGeometry padding;
+  final Widget separator;
   final Widget Function(BuildContext, T) itemBuilder;
   final ScrollMoreController controller;
   final Widget topWidget;
   final Widget bottomWidget;
 
-  final bool primary;
-  final double crossAxisSpacing;
-  final double mainAxisSpacing;
-  final int crossAxisCount;
-  final StaggeredTile Function(int) staggeredTileBuilder;
-
   @override
-  _RefreshableStaggeredGridViewState<T> createState() => _RefreshableStaggeredGridViewState<T>();
+  _RefreshableListViewState<T> createState() => _RefreshableListViewState<T>();
 }
 
-class _RefreshableStaggeredGridViewState<T> extends State<RefreshableStaggeredGridView<T>>
-    with AutomaticKeepAliveClientMixin<RefreshableStaggeredGridView<T>> {
+class _RefreshableListViewState<T> extends State<RefreshableListView<T>>
+    with AutomaticKeepAliveClientMixin<RefreshableListView<T>> {
   @override
   bool get wantKeepAlive => true;
 
+  // copy of widget.controller or new controller
   ScrollMoreController _controller;
   GlobalKey<RefreshIndicatorState> _refreshIndicatorKey;
-  bool _loading = false;
+
+  // loading, error message
+  bool _loading = true;
   String _errorMessage;
 
   @override
@@ -78,26 +69,26 @@ class _RefreshableStaggeredGridViewState<T> extends State<RefreshableStaggeredGr
   }
 
   Future<void> _getData() async {
+    // start refresh
     final func = widget.getData();
     _loading = true;
-    if (mounted) {
-      setState(() {});
-    }
+    if (mounted) setState(() {});
+
     return func.then((List<T> list) async {
+      // success to get data,  empty errorMessage
       _errorMessage = null;
       widget.data.clear();
-      if (mounted) {
-        setState(() {});
-      }
-      await Future.delayed(Duration(milliseconds: 20));
+      if (mounted) setState(() {});
+
+      await Future.delayed(Duration(milliseconds: 20)); // must delayed
       widget.data.addAll(list);
     }).catchError((e) {
+      // error arowsed
       _errorMessage = e.toString();
     }).whenComplete(() {
+      // finish loading
       _loading = false;
-      if (mounted) {
-        setState(() {});
-      }
+      if (mounted) setState(() {});
     });
   }
 
@@ -107,7 +98,7 @@ class _RefreshableStaggeredGridViewState<T> extends State<RefreshableStaggeredGr
     return RefreshIndicator(
       key: _refreshIndicatorKey,
       onRefresh: () => _getData(),
-      child: ListPlaceHolderText.from(
+      child: ListPlaceholderText.from(
         placeholderText: widget.placeholderText,
         onRefresh: _refreshIndicatorKey.currentState?.show,
         isLoading: _loading,
@@ -119,16 +110,12 @@ class _RefreshableStaggeredGridViewState<T> extends State<RefreshableStaggeredGr
             widget.topWidget ?? SizedBox(height: 0),
             Expanded(
               child: Scrollbar(
-                child: StaggeredGridView.countBuilder(
+                child: ListView.separated(
                   controller: _controller,
                   physics: AlwaysScrollableScrollPhysics(),
                   padding: widget.padding,
-                  primary: widget.primary,
-                  crossAxisSpacing: widget.crossAxisSpacing,
-                  mainAxisSpacing: widget.mainAxisSpacing,
-                  crossAxisCount: widget.crossAxisCount,
-                  staggeredTileBuilder: widget.staggeredTileBuilder,
                   itemCount: widget.data.length,
+                  separatorBuilder: (c, idx) => widget.separator ?? SizedBox(height: 0),
                   itemBuilder: (c, idx) => widget.itemBuilder(c, widget.data[idx]),
                 ),
               ),