+++ test/flutter_ahlib_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_ahlib/common/action_controller.dart';
+import 'package:flutter_ahlib/src/common/action_controller.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {