+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [1.0.1] - Initial flutter_ahlib
 
 + Initial `flutter_ahlib`
-+ Add `AppendIndicator`, `ListPlaceHolderText`, `AppendableListView`, `RefreshableListView`, `ScrollFloatingActionButton` widgets
++ Add `AppendIndicator`, `ListPlaceholderText`, `AppendableListView`, `RefreshableListView`, `ScrollFloatingActionButton` widgets
 + Add `ScrollMoreController`, `ScrollFabController` controllers
 
 ## [1.0.2] - Add widgets from mmnj
@@ -14,12 +14,20 @@
 
 + Fix `ScrollMoreController`.`scrollWithAnimate`
 
-## [1.0.5] - Update ListPlaceHolderText
+## [1.0.5] - Update ListPlaceholderText
 
-+ Add `ListPlaceHolderSetting` to list widgets
++ Add `ListPlaceholderSetting` to list widgets
 
 ## [1.0.8] - Add DrawerListView
 
 + Add `DrawerListView<T>` and some types
 + Rename `PopupMenuItem` to `PopupActionItem`
 + Update `RippleSizedView` actions
+
+## [1.0.9] - Update structure
+
++ Move all source file to `lib/src/` and export it on `lib/flutter_ahlib.dart`
++ Extract `lib/src/list/` package types to `lib/src/list/type.dart`
++ Rename `PlaceHolder` to `Placeholder`
++ Split `lib/src/list` to `lib/src/fab`
++ Add some document comments